requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
torch
pillow
torchvision



